restructured action creators and split them into counter, and results.  exported all of them via index.js within the actions folder
