+++ src/containers/Counter/Counter.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import * as actionCreators from "../../store/actions/actions";
+import * as actionCreators from "../../store/actions/index";
 import { connect } from "react-redux";
 
 import CounterControl from "../../components/CounterControl/CounterControl";
