+++ src/store/actions/actionTypes.js
@@ -1,0 +1,6 @@
+export const INCREMENT = "INCREMENT";
+export const DECREMENT = "DECREMENT";
+export const ADD_COUNTER = "ADD_COUNTER";
+export const MINUS_COUNTER = "MINUS_COUNTER";
+export const STORE_RESULT = "STORE_RESULT";
+export const DELETE_RESULT = "DELETE_RESULT";
