+++ src/store/actions/index.js
@@ -1,0 +1,2 @@
+export { add_counter, minus_counter, increment, decrement } from "./counter";
+export { store_result, delete_result } from "./result";
