+++ src/store/actions/result.js
@@ -1,0 +1,23 @@
+import * as actionTypes from "./actionTypes";
+
+export const saveResult = res => {
+  return {
+    type: actionTypes.STORE_RESULT,
+    result: res
+  };
+};
+
+export const store_result = res => {
+  return dispatch => {
+    setTimeout(() => {
+      dispatch(saveResult(res));
+    }, 2000);
+  };
+};
+
+export const delete_result = resultElId => {
+  return {
+    type: actionTypes.DELETE_RESULT,
+    resultElId: resultElId
+  };
+};
